+++ OgreMain/src/OgreProgressiveMesh.cpp
@@ -112,11 +112,11 @@ namespace Ogre {
             {
 			    if (quota == VRQ_PROPORTIONAL)
 			    {
-				    numCollapses = numVerts * reductionValue;
+				    numCollapses = static_cast<size_t>(numVerts * reductionValue);
 			    }
 			    else 
 			    {
-				    numCollapses = reductionValue;
+				    numCollapses = static_cast<size_t>(reductionValue);
 			    }
                 // Minimum 3 verts!
                 if ( (numVerts - numCollapses) < 3)
